Fix entities module dependencies to main common file
Move all identifiers used in entities/* from common.py to
entities/helper.py
Rationale: Sub-packages should not depend on files of the parent
hierarchy.
Previously, files in entities used absolute import "common" which is
ambiguous in some contexts.
