+++ common.py
@@ -6,7 +6,7 @@ import collections
 import csv
 import re
 from datetime import datetime
-from os import makedirs, path
+from os import path
 
 from typing import Any, Dict, Iterable, List, OrderedDict, Union
 
@@ -52,29 +52,6 @@ def now() -> str:
     return datetime.utcnow().replace(microsecond=0).isoformat() + 'Z'
 
 
-def escape_string(string: str) -> str:
-    """
-    Escape a string according to LibrePCB S-Expression escaping rules.
-    """
-    for search, replacement in STRING_ESCAPE_SEQUENCES:
-        string = string.replace(search, replacement)
-    return string
-
-
-def format_float(number: float) -> str:
-    """
-    Format a float according to LibrePCB normalization rules.
-    """
-    formatted = '{:.3f}'.format(number)
-    if formatted == '-0.000':
-        return '0.0'  # Remove useless sign
-    if formatted[-1] == '0':
-        if formatted[-2] == '0':
-            return formatted[:-2]
-        return formatted[:-1]
-    return formatted
-
-
 def format_ipc_dimension(number: float, decimal_places: int = 2) -> str:
     """
     Format a dimension (e.g. lead span or height) according to IPC rules.
@@ -89,13 +66,6 @@ def format_ipc_dimension(number: float, decimal_places: int = 2) -> str:
     # Note: Round to 1nm before truncating to avoid wrong results due to
     # inaccurate calculations leading in numbers like 0.79999999999999.
     return str(int(round(number, 6 - decimal_places)))
-
-
-def indent(level: int, lines: Iterable[str]) -> List[str]:
-    """
-    Indent the lines by the specified level.
-    """
-    return [' ' * level + line for line in lines]
 
 
 def sign(val: Union[int, float]) -> int:
@@ -135,19 +105,3 @@ def human_sort_key(key: str) -> List[Any]:
         return int(text) if text.isdigit() else text
 
     return [_convert(x) for x in re.split(r'(\d+)', key) if x]
-
-
-def serialize_common(
-    serializable: Any, output_directory: str, uuid: str, long_type: str, short_type: str
-) -> None:
-    """
-    Centralized serialize() implementation shared between Component, Symbol, Device, Package
-    """
-    dir_path = path.join(output_directory, uuid)
-    if not (path.exists(dir_path) and path.isdir(dir_path)):
-        makedirs(dir_path)
-    with open(path.join(dir_path, f'.librepcb-{short_type}'), 'w', newline='\n') as f:
-        f.write('2\n')
-    with open(path.join(dir_path, f'{long_type}.lp'), 'w', newline='\n') as f:
-        f.write(str(serializable))
-        f.write('\n')
