+++ entities/component.py
@@ -1,6 +1,6 @@
 from typing import Iterable, List, Optional
 
-from common import serialize_common
+from .helper import serialize_common
 
 from .common import (
     Author,
