+++ entities/package.py
@@ -1,6 +1,6 @@
 from typing import Iterable, List, Optional, Union
 
-from common import format_float, serialize_common
+from .helper import format_float, serialize_common
 
 from .common import (
     Align,
